Before merge, all tests except attribute working

+++ tests/correct2/recursive_def.py
@@ -1,4 +1,4 @@
-def fact(n::int): 
+def fact(n): 
     if n == 0: 
         return 1
     else:
